refactor: move cluster-wide tests into one dir, reorg, remove unused utils.
