+++ e2e_tests/Makefile
@@ -1,7 +1,7 @@
 .PHONY: fmt check test
 
 fmt:
-	isort \
+	isort -y \
 		-s ./tests/fixtures/pytorch_amp/model_def.py \
 		-s ./tests/fixtures/pytorch_amp/layers.py \
 		-s ./tests/fixtures/pytorch_amp/data.py \
