--- e2e_tests/tests/cluster/__init__.py
@@ -1,11 +1,0 @@
-from .cluster import (
-    cluster_slots,
-    det_version,
-    gpu_slots_per_agent,
-    max_slots_per_agent,
-    num_agents,
-    num_free_slots,
-    num_slots,
-    running_on_gpu,
-    wait_for_agents,
-)
